refactor to align APIs
exported commands become publish.map, publish.app and
publish.dependency. CLI commands become map, publish and dependency
