+++ classes/index.js
@@ -2,20 +2,18 @@
 
 const init = require('./init');
 const version = require('./version');
-const mapUpload = require('./map/upload');
+const alias = require('./alias');
+const publishMap = require('./publish/map');
 const publishApp = require('./publish/app');
 const publishDependency = require('./publish/dependency');
-const alias = require('./alias');
 
 module.exports = {
     init,
     version,
     alias,
-    map: {
-        upload: mapUpload
-    },
     publish: {
         app: publishApp,
+        map: publishMap,
         dependency: publishDependency
     }
 };
