+++ classes/publish/map.js
