+++ commands/publish.js
@@ -2,30 +2,18 @@
 
 const ora = require('ora');
 const PublishApp = require('../classes/publish/app');
-const PublishDependency = require('../classes/publish/dependency');
 const { resolvePath, logger } = require('../utils');
 
 const assetsPath = resolvePath('./assets.json').pathname;
 const assets = require(assetsPath);
 
-exports.command = 'publish [name] [version]';
+exports.command = 'publish';
 
-exports.aliases = ['p'];
+exports.aliases = ['p', 'pub'];
 
-exports.describe = `Publish an apps dependencies based on local assets.json file or publish an NPM package by given name and version.`;
+exports.describe = `Publish an apps dependencies based on local assets.json file.`;
 
 exports.builder = yargs => {
-    yargs
-        .positional('name', {
-            describe: 'Optional NPM package name.',
-            type: 'string'
-        })
-        .positional('version', {
-            describe:
-                'Optional NPM package version. Must be provided if "name" is given.',
-            type: 'string'
-        });
-
     yargs.options({
         server: {
             alias: 's',
@@ -63,6 +51,14 @@ exports.builder = yargs => {
             describe:
                 'Specify the path on local disk to CSS client side assets relative to the current working directory.',
             default: assets.css.input
+        },
+        name: {
+            describe: 'Specify the app name.',
+            default: assets.name
+        },
+        version: {
+            describe: 'Specify the app version.',
+            default: assets.version
         }
     });
 };
@@ -73,15 +69,7 @@ exports.handler = async function(argv) {
 
     try {
         const options = { logger: logger(spinner), ...argv };
-        if (argv.name && argv.version) {
-            success = await new PublishDependency(options).run();
-        } else {
-            success = await new PublishApp({
-                name: assets.name,
-                version: assets.version,
-                ...options
-            }).run();
-        }
+        success = await new PublishApp(options).run();
     } catch (err) {
         spinner.warn(err.message);
     }
